requirements.txt
# Core dependencies
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0

# Movie data
tmdbv3api==1.9.0

# Voice generation
elevenlabs==0.2.24

# Video processing
moviepy==1.0.3
opencv-python==4.8.1.78

# YouTube API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Caption generation
pysrt==1.1.2

# Image/video APIs
Pillow==10.1.0 